Implements the contexts for rendering the ui data
