+++ src/pages/Monitoring.tsx
@@ -6,8 +6,7 @@ import { SideBar } from "pages/monitoring/layout/SideBar";
 import "pages/monitoring/styles/monitoring.css";
 import type { UiManager } from "core/layout/MainLayout";
 import { LayoutUpdated } from "core/layout/mainLayout/hooks/layoutReducer";
-
-import { UserInMonitoringCTX } from "pages/monitoring/hooks/useUserInitiatives";
+import { UserInMonitoringCTX } from "pages/monitoring/hooks/useUserInitiativesCTX";
 
 export function Monitoring() {
   const { layoutDispatch } = useOutletContext<UiManager>();
